240507 임재현 DB 구조 변경 사항 적용

+++ server/src/main/java/org/web/application/server/entity/HeightEntity.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class HeightEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long heightId;
+    private Integer heightId;
 
     private Integer height;
